+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## Unreleased
-- ...
+- Add: message when user tries to enter fewer than minimum amount of points
 
 ## v0.1.0
 - Initial Release